[#12] chore: repository 일괄 변경
- *JpaRepository를 Repository 변경
