--- src/main/java/com/masterpiece/IPiece/common/domain/infra/DisclosureRepository.java
@@ -1,7 +1,0 @@
-package com.masterpiece.IPiece.common.domain.infra;
-
-import com.masterpiece.IPiece.common.domain.product.Disclosure;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DisclosureRepository extends JpaRepository<Disclosure, Long> {
-}
