+++ src/main/java/com/masterpiece/IPiece/market/infra/adapter/ProductQueryAdapter.java
@@ -1,8 +1,8 @@
 package com.masterpiece.IPiece.market.infra.adapter;
 
+import com.masterpiece.IPiece.common.domain.infra.ProductRepository;
 import com.masterpiece.IPiece.common.domain.product.Product;
 import com.masterpiece.IPiece.market.application.port.ProductQueryPort;
-import com.masterpiece.IPiece.market.infra.jpa.ProductJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,16 +13,16 @@ import java.util.Optional;
 @Component
 @RequiredArgsConstructor
 public class ProductQueryAdapter implements ProductQueryPort {
-    private final ProductJpaRepository repo;
+    private final ProductRepository productRepository;
 
     @Override
     public Page<Product> findActiveProducts(Pageable pageable) {
         // (이전) ACTIVE -> (변경) TRADE 기준으로 조회
-        return repo.findTradeProducts(pageable);
+        return productRepository.findTradeProducts(pageable);
     }
 
     @Override
     public Optional<Product> findById(Long productId) {
-        return repo.findById(productId);
+        return productRepository.findById(productId);
     }
 }
