+++ src/main/java/com/masterpiece/IPiece/market/infra/adapter/TradingInfoQueryAdapter.java
@@ -3,8 +3,8 @@ package com.masterpiece.IPiece.market.infra.adapter;
 import com.masterpiece.IPiece.common.domain.product.Disclosure;
 import com.masterpiece.IPiece.common.domain.product.ProductTradingInfo;
 import com.masterpiece.IPiece.market.application.port.TradingInfoQueryPort;
-import com.masterpiece.IPiece.market.infra.jpa.DisclosureJpaRepository;
-import com.masterpiece.IPiece.market.infra.jpa.ProductTradingInfoJpaRepository;
+import com.masterpiece.IPiece.market.infra.jpa.DisclosureRepository;
+import com.masterpiece.IPiece.market.infra.jpa.ProductTradingInfoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Component;
@@ -16,17 +16,17 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TradingInfoQueryAdapter implements TradingInfoQueryPort {
 
-    private final ProductTradingInfoJpaRepository tradingInfoRepo;
-    private final DisclosureJpaRepository disclosureRepo;
+    private final ProductTradingInfoRepository productTradingInfoRepository;
+    private final DisclosureRepository disclosureRepository;
 
     @Override
     public Optional<ProductTradingInfo> findTradingInfo(Long productId) {
-        return tradingInfoRepo.findById(productId);
+        return productTradingInfoRepository.findById(productId);
     }
 
     @Override
     public List<Disclosure> findDisclosures(Long productId, int limit) {
-        return disclosureRepo.findByProductId(
+        return disclosureRepository.findByProductId(
                 productId,
                 PageRequest.of(0, Math.max(limit, 1))
         );
